requirements-test.txt
django-nose>=1.4.2,<2
mock>=1.3.0,<4
nose>=1.3.7,<2
tox>=2.2.1,<3
